Se terminaron varios mensajes, falta probarlos. Falta plato listo y consultar pedido.

+++ cliente/src/cliente.c
@@ -93,7 +93,7 @@ int main(int argc, char **argv) {
 
     		    		printf("Debe ingresar mas parámetros.\n");
 
-    		    		enviar_09_confirmar_pedido( g_ip_app, g_puerto_app,argv[3],argv[4]);
+    		    		enviar_09_confirmar_pedido( g_ip_app, g_puerto_app,argv[3],(uint32_t)argv[4]);
     		    		//enviar_09_confirmar_pedido_hack ( g_ip_app, g_puerto_app );
 
     		    		break;
